Renamed test runner visitor file
